Moves tests into more appropriately namespaced directories.
